fix(seo): server-render the pages Google files as duplicates
Search Console's indexing alert named three new reasons. They resolve to
one metadata bug and one architectural gap.

The bug: noindexMetadata() never declared `alternates`, and Next resolves
each metadata key against the nearest ancestor that declares it. Omitting
the key does not mean "no canonical" -- it means the parent segment's
canonical is inherited. 39 staff routes therefore answered `noindex`
while naming an indexable page as their canonical (/manual/logs pointed
at /manual, /manual/item/balance at /manual/item), which is the one
combination Google asks you not to ship: the noindex can be consolidated
onto the canonical target. Those targets are the best non-homepage URLs
on the site -- /manual/bloodline alone draws 7.7k impressions at
position 3.7. Declaring the key with a null value emits no tag.

The gap: forum threads, forum boards, profiles and concept art resolve
their content client-side, so the server sent every URL of each kind the
same shell. Profiles were 505 characters reading "Loading profile /
Fetching Public User Data"; threads were 2,389 characters of boilerplate
that is byte-identical across all 381 of them, with the title reaching
the crawler only inside <title>. Google read the first as soft 404s and
folded the rest together as duplicates.

Each of those routes already resolved the identifying row server-side to
build its metadata, so the fix is to hand that work to the client
component rather than throw it away:

- Threads and boards became server components that resolve the first page
  and seed the client query with it. Both call the same helper the tRPC
  procedure calls -- fetchForumThreadPage, extracted here, and the
  existing getInfiniteThreads -- so the seeded payload cannot drift from
  the fetched one. An unknown thread or board now answers 404 rather than
  a 200 carrying a "not found" body, matching /username/<name>.
- Profiles pass the row generateMetadata already loaded, and the
  placeholder renders it: the real name, level, rank, village and intro
  copy instead of grey bars. That also removes the shift the placeholder
  existed to absorb, since the panel no longer changes size.
- Concept art seeds prompt and creator. It deliberately does not 404 on a
  miss: indexableConceptArt excludes images that are still generating.

Finally, the wallpaper is the largest contentful paint on most pages and
Search Console reports every LCP group at 4.0s. It was already served in
three renditions, but eager and high priority only reorder it against
work the browser has found -- it cannot start before the parser reaches
the body. It is now preloaded from <head>, under media queries that are
mutually exclusive so no viewport pulls a second copy.

Verified against a local server: noindex routes emit no canonical, the
indexable ones keep theirs, profiles went 505 -> 3,125 characters with
the player's own name and village, boards and threads now carry their
titles and posts, and unknown ones answer 404.
